refactor: extract columns from runs data table
